Move parts to Search class
